refactor(scripts): repoint package, CI, and tests
Everything that named a moved script now names its new home: the
13 package.json script entries, the npm-publish workflow step, the
perf fixture import, and the test suite's imports and read paths.

Two path lookups were built segment by segment rather than as one
string, so a search for the old text missed them. The weekly update
chain resolved update-manifest.mts by joining 'scripts' and 'npm',
and three publish tests read their source files the same way. Both
now include the repo segment.
