+++ package.json
@@ -19,43 +19,43 @@
     "linkDirectory": false
   },
   "scripts": {
-    "analyze-ci-failures": "node scripts/testing/analyze-ci-failures.mts",
+    "analyze-ci-failures": "node scripts/repo/testing/analyze-ci-failures.mts",
     "build": "node scripts/repo/build.mts",
     "check": "node scripts/fleet/check.mts",
-    "check-trusted": "node scripts/npm/check-trusted-packages.mts",
+    "check-trusted": "node scripts/repo/npm/check-trusted-packages.mts",
     "check:paths": "node scripts/fleet/check/paths-are-canonical.mts",
     "clean": "node scripts/fleet/clean.mts",
     "cover": "node scripts/fleet/cover.mts",
     "fix": "node scripts/fleet/fix.mts",
     "format": "node scripts/fleet/format.mts",
     "format:check": "node scripts/fleet/format.mts --check",
-    "generate-actions-allow-list": "node scripts/ci/generate-actions-allow-list.mts",
-    "generate-actions-tree": "node scripts/ci/show-actions-tree.mts",
-    "inline-action-shas": "node scripts/ci/inline-action-versions-as-shas.mts",
+    "generate-actions-allow-list": "node scripts/repo/ci/generate-actions-allow-list.mts",
+    "generate-actions-tree": "node scripts/repo/ci/show-actions-tree.mts",
+    "inline-action-shas": "node scripts/repo/ci/inline-action-versions-as-shas.mts",
     "lint": "node scripts/fleet/lint.mts",
     "lockstep": "node scripts/fleet/lockstep.mts",
     "lockstep:emit-schema": "node scripts/fleet/lockstep-emit-schema.mts",
-    "make-npm-override": "node scripts/npm/make-npm-override.mts",
-    "npm:configure-staged": "node scripts/npm/configure-staged-publishing.mts",
+    "make-npm-override": "node scripts/repo/npm/make-npm-override.mts",
+    "npm:configure-staged": "node scripts/repo/npm/configure-staged-publishing.mts",
     "npm:publish": "node scripts/fleet/publish-pipeline.mts",
-    "npm:wire-port": "node scripts/npm/wire-npm-port.mts",
-    "package-npm-access": "node scripts/npm/set-npm-package-access.mts",
-    "package-npm-publish": "node scripts/npm/publish-npm-packages.mts",
+    "npm:wire-port": "node scripts/repo/npm/wire-npm-port.mts",
+    "package-npm-access": "node scripts/repo/npm/set-npm-package-access.mts",
+    "package-npm-publish": "node scripts/repo/npm/publish-npm-packages.mts",
     "perf": "node scripts/repo/perf.mts",
     "security": "node scripts/fleet/security.mts",
     "setup": "node scripts/repo/setup.mts",
     "prepare": "node scripts/fleet/install-git-hooks.mts && node scripts/fleet/prepare.mts",
     "prepublishOnly": "echo 'ERROR: Use GitHub Actions workflow for publishing' && exit 1",
-    "release-npm": "node scripts/npm/release-npm-packages.mts",
+    "release-npm": "node scripts/repo/npm/release-npm-packages.mts",
     "test": "node scripts/fleet/test.mts",
     "test:fuzz": "node scripts/repo/fuzz.mts",
-    "test:npm": "node scripts/npm/run-vitest-npm.mts",
-    "test:npm:floor": "node scripts/npm/floor-node-smoke.mts",
+    "test:npm": "node scripts/repo/npm/run-vitest-npm.mts",
+    "test:npm:floor": "node scripts/repo/npm/floor-node-smoke.mts",
     "type": "node node_modules/typescript/bin/tsc --noEmit -p .config/fleet/tsconfig.check.json",
     "update": "node scripts/fleet/update.mts",
     "update:external-tools": "node scripts/repo/update-external-tools.mts",
-    "validate-ci": "node scripts/testing/reproduce-ci-locally.mts",
-    "validate-packages": "node scripts/testing/validate-package-tests.mts",
+    "validate-ci": "node scripts/repo/testing/reproduce-ci-locally.mts",
+    "validate-packages": "node scripts/repo/testing/validate-package-tests.mts",
     "setup-security-tools": "node .claude/hooks/fleet/setup-security-tools/install.mts",
     "ci:local": "node scripts/fleet/agent-ci-skip-locks.mts run --all --quiet --pause-on-failure --github-token",
     "weekly-update": "node scripts/fleet/weekly-update.mts",
