+++ perf/npm/json-stable-stringify.perf.mts
@@ -7,7 +7,7 @@ import { getDefaultLogger } from '@socketsecurity/lib/logger/default'
 import fastJsonStableStringify from 'fast-json-stable-stringify'
 import { Bench } from 'tinybench'
 
-import { PERF_NPM_FIXTURES_PATH } from '../../scripts/constants/paths.mts'
+import { PERF_NPM_FIXTURES_PATH } from '../../scripts/repo/constants/paths.mts'
 
 const logger = getDefaultLogger()
